unidler: squash some stub files into proxier.go
