--- pkg/network/proxy/unidler/proxier_openshift.go
@@ -1,18 +1,0 @@
-package unidler
-
-// Some extra hacking for openshift-specific stuff
-
-import "k8s.io/kubernetes/pkg/util/async"
-
-func (p *Proxier) SyncProxyRules() {
-	p.syncProxyRules()
-}
-
-func (p *Proxier) SetSyncRunner(b *async.BoundedFrequencyRunner) {
-	p.syncRunner = b
-}
-
-func (p *Proxier) ReloadIPTables() {
-	p.forceReload = true
-	p.syncProxyRules()
-}
